+++ packages/ibm-products/src/components/Datagrid/Datagrid.docs-page.js
@@ -148,13 +148,14 @@ useDatagrid({
   data,
   onColResizeEnd: (currentColumn, allColumns) =>
       console.log(currentColumn, allColumns),
+  resizerAriaLabel: 'Resize column',
 });
           `,
         },
       },
       {
         description:
-          'Disabling the resizable columns is possible by specifying `disableResizing: true` within the `useDatagrid` hook.',
+          'Disabling the resizable columns is possible by specifying `disableResizing: true` within the `useDatagrid` hook. To pass in your own translated label for the column resizer, add the `resizerAriaLabel` property',
         source: {
           language: 'jsx',
           code: `
@@ -162,6 +163,7 @@ useDatagrid({
   columns,
   data,
   disableResizing: true,
+  resizerAriaLabel: 'Resize column',
 });
           `,
         },
@@ -361,6 +363,7 @@ const datagridState = useDatagrid(
     columns,
     data,
     onRowSelect: (row, event) => console.log(row, event),
+    batchActionMenuButtonLabel: 'More',
   },
   useSelectRows
 );