+++ packages/ibm-products/src/components/EmptyStates/assets/NotFoundIllustration.js
@@ -17,7 +17,7 @@ import uuidv4 from '../../../global/js/utils/uuidv4';
 // The block part of our conventional BEM class names (blockClass__E--M).
 const blockClass = `${pkg.prefix}--empty-state`;
 
-const NotFoundIllustration = ({ title, theme, size, ...rest }) => {
+export const NotFoundIllustration = ({ theme, title, size, ...rest }) => {
   const svgId = uuidv4();
 
   return (